% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\docType{package}
\name{auto}
\alias{auto}
\alias{auto-package}
\title{Run Automatic Package Checks}
\description{
A personal package to automate the checks I think are
  important when reviewing R package Application papers for Methods in
  Ecology and Evolution. It isn't not designed to be configurable, nor
  intended as a statement of how all R packages should be implemented.

  The only exported function is \code{\link{check}}
}
