% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.R
\name{check}
\alias{check}
\title{Run Automatic Package Checks}
\usage{
check(pkg, dir = NULL)
}
\arguments{
\item{pkg}{a string giving the directory of the github repo in the format
"user/repo"}

\item{dir}{an optional file path giving a directory into which the package
source code will be saved. If \code{NULL}, a temporary directory will be
used.}
}
\description{
Run some quick automatic checks on a package hosted on GitHub,
  using \code{\link[devtools:check]{devtools::check}()} and
  \code{\link[goodpractice:gp]{goodpractice::gp}()}.
}
\details{
The results of \code{devtools::check()} and
  \code{goodpractice::gp()} will be printed to the console. RStudio markers
  are suppressed, and some checks are skipped.

  To check a package hosted on CRAN, you can use the cran github mirror,
  e.g.: "cran/devtools".
}
\examples{
\dontrun{
auto::check("goldingn/auto")
}
}
